Simplified and split out effects to their respective modules
